% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gapminder.R
\docType{data}
\name{my_gapminder}
\alias{my_gapminder}
\title{An excerpt of the data available at Gapminder.org.}
\format{a data frame with 1704 observations and 6 variables:
\describe{
\item{country}{names of countries}
\item{continent}{continents where the countries are in}
\item{year}{year of the data, in increment of 5 years}
\item{lifeExp}{life expectancy at birth, in years}
\item{pop}{population}
\item{gdpPercap}{GDP per capita, inflation-adjusted, in USD}
}}
\source{
\url{"http://www.gapminder.org/data/"}
}
\usage{
my_gapminder
}
\description{
The famous \code{\link[datasets]{gapminder}} dataset taken from the
\code{datasets} package to practice adding datasets to R package.
}
\keyword{datasets}
