% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-validation function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input used as the number of folds.}
}
\value{
Numeric indicating the average mean square error across \code{k}
folds
}
\description{
This function performs a random forest cross validation on the penguin data.
}
\examples{
my_rf_cv(5)

}
\keyword{inference}
