% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector input to be used as the data for the t-test.}

\item{alternative}{String input that indicates the alternative hypothesis,
which must be one of  \code{"two.sided"}, \code{"greater"}, or
\code{"less"}.}

\item{mu}{Numeric input to be used as the true value of mean.}
}
\value{
a list having the following elements, \code{test_stat} the numeric
test statistic, \code{df} the degrees of freedom, \code{alternative} the
string indicating the alternative hypothesis, \code{p_val} the numeric
p-value.
}
\description{
This function performs a one sample t-test.
}
\examples{
x <- rnorm(10, mean = 0, sd = 1)
my_t.test(x, "two.sided", 0)
my_t.test(x, mu = -1, alternative = "less")

}
\keyword{inference}
