% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{T-test Function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{Data vector on which t-test is performed.}

\item{alternative}{A String (less, greater, or two.sided),
against which the null-hypothesis will be compared.}

\item{mu}{A numeric value representing the null hypothesis for the t-test.}
}
\value{
List containing the numeric test statistic, numeric degrees of freedom,
string representing the type of t-test performed, and the numeric p-value.
}
\description{
This function performs a t-test on a data vector.
}
\examples{
my_t_test(my_gapminder$lifeExp, "greater", 60)
my_t_test(my_gapminder$lifeExp, "two.sided", 60)

}
\keyword{inference}
