% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data frame input.}

\item{cl}{Vector inpur containing the true class of the training data.}

\item{k_nn}{Numeric input of the k-nearest neighbors.}

\item{k_cv}{Numeric input of number of folds.}
}
\value{
List output containing predicted class and cross validation error.
}
\description{
This function ____.
}
\examples{
library(palmerpenguins)
data(package = "palmerpenguins")
penguins_omit <- na.omit(penguins)
my_knn_cv(train = penguins_omit["bill_length_mm"], cl = penguins_omit["species"],
  k_nn = 1, k_cv = 5)

}
