% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k nearest neighbors cross validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{A data frame to predict classifications with.}

\item{cl}{The true classifications of each observation in \code{train}.}

\item{k_nn}{A numeric representing the number of neighbors to use.}

\item{k_cv}{A numeric representing the number of folds to use in cross validation.}
}
\value{
A list with two elements:
\code{"class"}: the predicted class of each element of \code{train},
\code{"cv_error"}: the cross validation error for the predictions.
}
\description{
This function predicts categories based on the k nearest neighbors and uses
cross validation to indicate efficacy.
}
\examples{
# I have spent an hour on this and can't get a working example

}
