% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K Nearest Neighbors Cross Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{The dataset on which your model will be trained.}

\item{cl}{The vector containing the actual classes of your observations.}

\item{k_nn}{The number of nearest neighbors used to predict the class of an observation.}

\item{k_cv}{The number of folds your training data will be divided into for cross-validation.}
}
\value{
A list containing vector "class", which contains the predicted classes using
the entire data as training data, and the numeric "cv_err" which is the mean
cross-validation error for your chosen "k".
}
\description{
Function performs cross validation to determine appropriate "k"
for k-nearest-neighbors.
}
\examples{
library(dplyr)
library(class)
my_knn_cv(my_penguins, my_penguins$species, 1, 5)
my_knn_cv(my_penguins, my_penguins$species, 5, 5)

}
\keyword{prediction}
