% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-nearest neighbors with cross validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data frame used for training the model.}

\item{cl}{Vector of true classifications for entries in \code{train}.}

\item{k_nn}{Integer indicating how many neighbors should be considered in the
k-nearest neighbors algorithm.}

\item{k_cv}{Integer indicating how many folds \code{train} will be divided
into for k-fold cross-validation.}
}
\value{
A list containing a vector of classifications predicted by the model
using \code{train} as both the training data and the testing data and a
numeric representing the cross-validation estimate from the k-fold
cross-validation.
}
\description{
Builds a classification model for data \code{train} using the k-nearest
neighbors algorithm with k = \code{k_nn} and then applies k-fold
cross-validation with k = \code{k_cv}.
}
\examples{
train <- na.omit(my_penguins)[,3:6]
cl <- na.omit(my_penguins)[1]
my_knn_cv(train, cl, 1, 5)
my_knn_cv(train, cl, 5, 5)

}
\keyword{prediction}
