% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-nearest neighbors function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Dataframe of training data}

\item{cl}{Vector input of true class value of the training data}

\item{k_nn}{Numeric number of neighbors used in k-nearest neighbors}

\item{k_cv}{Numeric number of folds used in k-fold cross-validation}
}
\value{
List object containing the predicted class value of the data, the CV
misclassification rate and the training misclassification rate
}
\description{
This function predicts the class value by k-fold Cross-validation and k-nearest neighbors algorithms.
}
\examples{
my_knn_cv(my_gapminder[, 4:5], continent, 1, 5)

}
\keyword{prediction}
