% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{My k-nearest-neighbors cross-validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{A data frame of the training data.}

\item{cl}{A vector of strings containing the true classes of each observation
in \code{train}.}

\item{k_nn}{A numeric specifying the number of neighbors to be used in the
k-nearest-neighbors algorithm.}

\item{k_cv}{A numeric specifying the number of folds to be used in the
cross-validation.}
}
\value{
A list with two elements: \code{class}, a vector of predicted classes
for each observation in \code{train} using \code{k_nn} nearest neighbors
and \code{k_cv} cross-validation folds and \code{cv_error}, the average
misclassification rate of the k-nearest-neighbors algorithm across all
the folds in the cross-validation.
}
\description{
This function runs a k-nearest-neighbors classification algorithm on a
specified set of data and performs k-fold cross-validation on it.
}
\examples{

my_train <- my_penguins \%>\% dplyr::select(body_mass_g,
                                         flipper_length_mm,
                                         bill_length_mm,
                                         bill_depth_mm)
my_cl <- my_penguins$species
my_knn_cv(train = my_train,
          cl = my_cl,
          k_nn = 3,
          k_cv = 5)

}
\keyword{prediction}
