% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame to be trained.}

\item{cl}{Vector of true class value of the training data.}

\item{k_nn}{Numeric input representing number of neighbors.}

\item{k_cv}{Numeric input representing number of folds.}
}
\value{
A list with a vector of the predicted class and cross-validation
misclassification error.
}
\description{
This function performs k-nearest neighbors cross-validation.
}
\examples{
train <- rbind(iris3[, , 1], iris3[, , 2], iris3[, , 3])
cl <- factor(c(rep("s", 50), rep("c", 50), rep("v", 50)))
my_knn_cv(train, cl, k_nn = 1, k_cv = 5)
my_knn_cv(train, cl, k_nn = 5, k_cv = 5)

}
\keyword{prediction}
