% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula input of an output class and predictor variables
from \code{data}.}

\item{data}{Data frame input.}
}
\value{
Matrix displaying the coefficient, standard error, observed
t value, and p_value for the linear model.
}
\description{
This function ____.
}
\examples{
data(mtcars)
my_lm(formula = mpg ~ hp + wt, data = mtcars)

}
