% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear fit generator}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula class object to fit \code{data} to.}

\item{data}{Data frame that will be fit onto \code{formula}.}
}
\value{
Table of coefficients generated through the linear fit. Contains the
estimates of the coefficients of the variables provided in the formula
(and the intercept), the standard error of these estimates, the
corresponding t values, and the p values of these t values for each
coefficient.
}
\description{
Performs a linear fit on data given a formula.
}
\examples{
x <- c(0.5, 0.8, 1.5, 0.5, 0.2)
y <- c(1.0, 1.2, 0.1, 0.2, 1.4)
df <- cbind(data.frame(y), data.frame(x))
my_lm(y ~ x, df)

}
\keyword{inference}
