% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{The linear regression model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{an object class as a symbolic description of the
model to be fitted.}

\item{data}{input data frame.}
}
\value{
a table similar to coefficient table of lm()
}
\description{
This function is used to fit linear model and carry out linear regression.
}
\examples{
formula <- mpg ~ hp + wt
data <- mtcars
my_lm(formula, data)


}
\keyword{inference,}
\keyword{linear}
\keyword{model,}
\keyword{prediction.}
\keyword{regression,}
