% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{The formula object to be tested}

\item{data}{a numeric data frame representing a data set with the variables in
\code{formula}.}
}
\value{
a list representing the linear model for \code{data} with the given
\code{formula} containing:
Estimate: the estimated coefficients for \code{formula},
Std. Error: the average error of the predicted model,
t value: a numeric representing the t-value of each estimated coefficient,
Pr(>|t|) the probability of a more extreme estimated coefficient.
}
\description{
This function calculates a regression model.
}
\examples{
y <- c(2, 5, 6, 8, 9)
x <- c(1, 2, 3, 4, 5)
example.df <- data.frame("y" = y, "x" = x)
my_lm(y ~ x, example.df)
my_lm(x ~ y, example.df)

}
