% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Modeling Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula specifying which variable is being predicted by
which other variables.}

\item{data}{A set of data from which the model will be built.}
}
\value{
A table displaying the linear coefficients, the standard error
of those coefficients, the t-value, and the p-value for each variable.
}
\description{
This function applies a linear model to a set of data, and returns the results.
}
\keyword{inference}
\keyword{prediction}
