% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{linear model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Numeric input of a formula of the response variable and explainary variable}

\item{data}{Numeric dataframe containing the data that is used in \code{formula}}
}
\value{
Table object containing estimated coefficients, standard error, t value, and p value
}
\description{
This function fits a linear model to input data.
}
\examples{
my_lm(lifeExp~gdpPercap+continent, my_gapminder)

}
\keyword{inference}
