% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{My Linear Model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula object with the linear regression equation to be
calculated from data set \code{data}.}

\item{data}{A data frame containing the data to be modeled.}
}
\value{
A table with coefficient estimates and standard errors for each
variable specified by \code{formula} in the linear model as well as
t-values and p-values for two-sided t-tests performed on each coefficient
estimate.
}
\description{
This function calculates a linear regression model for given data.
}
\keyword{inference}
\keyword{prediction}
