% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object.}

\item{data}{Input data frame.}
}
\value{
A table containing the estimate, standard error, t value,
and Pr(>|t|).
}
\description{
This function fits a linear model.
}
\examples{
my_lm(formula = mpg ~ hp + wt, data = mtcars)
my_lm(formula = mpg ~ hp * wt, data = mtcars)

}
