% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input of the number of folds.}
}
\value{
Numeric output of the cross-validation error.
}
\description{
This function predicts a variable using covariates from the same data set.
}
\examples{
my_rf_cv(k = 5)

}
\keyword{prediction}
