% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric denoting number of folds to use in the cross validation.}
}
\value{
Numeric that is the mean squared error of predicted values compared
to true body mass.
}
\description{
Performs random forest cross validation on the "penguins" data set,
predicting "body_mass_g" with covariates "bill_length_mm", "bill_depth_mm",
and "flipper_length_mm", with 100 trees.
}
\examples{
my_rf_cv(5)
my_rf_cv(30)

}
\keyword{prediction}
