% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random-Forest Cross Validation method}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{the number of folds}
}
\value{
numeric with cross validation error
}
\description{
This function predict output \code{body_mass_g} using covariates \code{bill_length_mm},
\code{bill_depth_mm}, and \code{flipper_length_mm}.
}
\examples{
my_rf_cv(5)

}
\keyword{cross}
\keyword{prediction,}
\keyword{validation}
