% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{random forest cross validation function for the palmerpenguins penguins data}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric representing the number of folds to use in cross validation.}
}
\value{
a numeric representing the cross validation error.
}
\description{
This functions determines the cross validation error for modeling body_mass_g
using bill_depth_mm, bill_length_mm, and flipper_length_mm with random trees.
}
