% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross Validation Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{A number representing the number of folds the data will be divided into.}
}
\value{
A number representing the mean cross-validation standard error for that
number of folds.
}
\description{
This function performs cross validation for values of k in random forest.
}
\examples{
library(randomForest)
library(dplyr)
my_rf_cv(2)
my_rf_cv(10)

}
\keyword{prediction}
