% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{random forest with k-fold cross validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Integer indicating how many folds \code{train} will be divided
into for k-fold cross-validation.}
}
\value{
numeric representing the cross-validation error from the k-fold
cross-validation.
}
\description{
Builds a classification model for the \code{my_penguins} data set to predict
\code{body_mass_g} based on \code{bill_length_mm}, \code{bill_depth_mm}, and
\code{flipper_length_mm} and applies k-fold cross-validation with
k = \code{k}.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
