% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{random forest function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric number of folds used in k-fold cross-validation}
}
\value{
Numeric number of the CV mean squared error
}
\description{
This function predicts the lifeExp in \code{my_gapminder} dataset by gdpPercap
with k-fold Cross-validation and random forest algorithms.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
