% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{My random forest cross validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{A numeric containing the number of folds to use in the k-fold
cross-validation.}
}
\value{
A numeric representing the average mean-squared-error across all
\code{k} folds.
}
\description{
This function calculates the cross-validation error of a k-fold
cross-validation of a 100-branch random forest algorithm performed on three variables
of the \code{penguins} data.
}
\examples{
my_rf_cv(3)

}
\keyword{prediction}
