% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input representing numbers of folds.}
}
\value{
Numeric indicating the cross-validation error.
}
\description{
This function performs random forest cross validation.
}
\examples{
my_rf_cv(5)
my_rf_cv(10)

}
