% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{Statistics t test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Vector of numerics representing sample data.}

\item{alternative}{String containing type of alternative hypothesis with
one of the following values "two.sided", "less", or "greater".}

\item{mu}{Numeric containing null hypothesis mean value.}
}
\value{
List containing "test_stat": the t statistic calculated, "df": the
degrees of freedom, "alternative": the value of \code{alternative},
"p_val": the p-value calculated from the t test.
}
\description{
This function performs a t test given data and a hypothesis.
}
\examples{
my_t.test(c(0.5, 0.6, 0, 1.5, 4), "two.sided", 2)
my_t.test(c(1.0, 1.0), "less", 0)

}
\keyword{inference}
