% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{the t-test calculator}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{the numeric vector of lifeExp pulled from the gapminder data}

\item{alternative}{the one-sided or two-sided test}

\item{mu}{the mean value tested given from the null hypothesis (mu = 60)}
}
\value{
A list indicating test statistics, p-value, degree of freedom,
and the conclusion on whether we reject or fail to reject the null hypothesis
tested from lifeExp data in the gapminder package.
}
\description{
This function calculates values associated to the t-test such as the
test statistics, p-value, degree of freedom, and the conclusion on
whether we reject or fail to reject the null hypothesis tested from
lifeExp data in the gapminder package.
}
\examples{
set.seed(302)
p <- 0.4
flip_coin <- rbinom(100, size = 1, prob = p)
my_t.test(as.numeric(flip_coin), alternative = "greater", mu = p)

}
\keyword{hypothesis,}
\keyword{inference}
