% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector representing a data set to conduct a t-test on.}

\item{alternative}{The alternative to be tested with: either \code{"greater"},
\code{"less"}, or \code{"two.sided"}.}

\item{mu}{A numeric representing the assumed mean of what \code{x} is a sample of.}
}
\value{
A list with 4 elements:
test-stat: a numeric representing the hypothesized t-score of sample \code{x},
df: a numeric representing the degrees of freedom of \code{x},
alternative: the passed in \code{alternative},
p-val: a numeric representing the probability of a sample as extreme as
\code{x} occurring.
}
\description{
This function runs a t-test on a set of data.
}
\examples{
 my_t.test(c(1, 2, 3, 4, 5), "two.sided", 0)
 my_t.test(x = 1:50, "greater", 10)

}
