% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{My t-test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data on which t-test is performed.}

\item{alternative}{String input that specifies the type of t-test to be
performed on \code{x}, must be either \code{"less"} or \code{"greater"}
for a one-sided t-test or \code{"two.sided"} for a two-sided test.}

\item{mu}{Numeric representing the null hypothesis value of the mean.}
}
\value{
List with four elements: \code{test_stat} is a numeric contianing
the value of the test statistic of \code{x}, \code{df} is a numeric
containing the degrees of freedom used in the t-test, \code{alternative}
is a string that represents the type of t-test used (the value of the
\code{alternative} parameter), and \code{p-val} is a numeric containing
the p-value for the t-test.
}
\description{
This function performs a t-test on a set of data against a hypothesis.
}
