% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-Test Function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data to be tested.}

\item{alternative}{Character string specifying the alternative hypothesis,
must be "two.sided", "less", or "greater".}

\item{mu}{Numeric input indicating the null hypothesis value of the mean.}
}
\value{
A list with the numeric test statistic, degree of freedom,
\code{alternative}, and the numeric p-value.
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t.test(1:10, alternative = "two.sided", mu = 5)
my_t.test(c(0, 0.5, -0.5, 0.2, 0), alternative = "less", mu = 0)

}
