% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{T Test Function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric input of a vector containing a set of data.}

\item{alternative}{String input indicating the type of hypothesis test.}

\item{mu}{Numeric input representing the value for the null hypothesis of
the mean, defaults to \code{0}.}
}
\value{
List output containing the t-statistic, df, \code{alternative},
and p-value.
}
\description{
This function ____.
}
\examples{
my_t_test(c(10, 15, 14, 16), alternative = "two.sided", mu = 13)

}
